+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+Unreleased
+-----------------
+
+### Changed
++ Query::getResult() renamed to Query::execute()
++ Query::prepareStatement() renamed to Query::perpare()
+
 0.7.1 [2016-11-14]
 -----------------